+++ 00_Uebung/02_Servo/README.md
@@ -1,0 +1,9 @@
+# Servo
+
+Dieses Programm lässt den ServoMotor hin und her schwenken. <br/>
+Das "ServoMotor.write(x); ist dafür verantwortlich, wie der Servo positioniert wird.<br/>
+Spiele den Code auf das Elektroauto und sieh, was passiert.<br/>
+Nachdem kannst du die Zahlen in der "ServoMotor.write(x);" ändern.<br/>
+Spiele den Code erneut rüber und siehe was passiert.<br/>
+
+
