--- 00_Uebung/Distanz_Serial_Monitor/README.md
@@ -1,5 +1,0 @@
-# Distanz Serial Monitor
-
-Nachdem du den Code auf das Arduino rübergespielt hast, drücke auf der Tastatur CTRL Shift und M gleichzeitig. <br/>
-Ein neues Fenster öffnet sich.<br/>
-Halte deine Hand vor den Ultraschallsensor. Was erkennst du auf dem Computer?
