+++ 00_Uebung/README.md
@@ -1,7 +1,4 @@
 # Übungen
 
 In diesem Ordner sind mehrere Übungen zu finden rund um das autonome Fahrzeug.
-Klicke auf die gewünschte Übung und lade alle Dateien im Ordner herunter. (Icon oben rechts).
-Die Datei README.md muss nicht heruntergeladen werden. 
-Optional kann der Code auch direkt kopiert werden.
-Jeder Übung ist die Aufgabe im Code wie auch auf GitHub beschrieben.
+
