+++ 03_Fertigung/README.md
@@ -2,4 +2,10 @@
 
 Hier sind Dateien zur Fertigung des autonomen Fahrzeugs zu finden.
 
-![Under Construction](https://navtec.de/wp-content/uploads/2018/02/under-construction-2408061_1920.png)
+</br>
+
+<h2>Die einzelnen Bauteile erklärt</h2>
+
+<h3>[DC-Motor](https://www.youtube.com/watch?v=6kJuuPxo4Q0)</h3>
+<h3>[Servo-Motor](https://www.youtube.com/watch?v=g4BM1pjm64s&t=3s)</h3>
+<h3>[Ultraschallsensoren](https://www.youtube.com/watch?v=BEDGEZsV7IY)</h3>
