--- 99_Solution/EAuto/MotorSteuerung.h
@@ -1,19 +1,0 @@
-#ifndef MOTORSTEUERUNG_H
-#define MOTORSTEUERUNG_H
-
-
-#include <Arduino.h>
-
-void InitialisierungMotor(void);
-
-
-void GeradeFahren(void);
-
-void LinksFahren(void);
-
-void RechtsFahren(void);
-
-void StoppFahren(void);
-
-
-#endif
