+++ 99_Solution/Loesung1/Loesung1.ino
@@ -16,6 +16,7 @@ Das Fahrzeug dreht sich dann ab , wo mehr Platz zur Verfügung steht.
 Sonstige Infos:
 
 Steuerung der Motoren erfolgt über die Datei MotorSteuerung.cpp. Die Header-Datei MotorSteuerung MotorSteurung.h dient als Zuordnung der Funktionen*/
+
 #include <Servo.h>
 #include <NewPing.h>
 #include "Motorsteuerung.h"
