+++ 99_Solution/Loesung1/MotorSteuerung.h
@@ -1,0 +1,19 @@
+#ifndef MOTORSTEUERUNG_H
+#define MOTORSTEUERUNG_H
+
+
+#include <Arduino.h>
+
+void InitialisierungMotor(void);
+
+
+void GeradeFahren(void);
+
+void LinksFahren(void);
+
+void RechtsFahren(void);
+
+void StoppFahren(void);
+
+
+#endif
