+++ 99_Solution/Loesung2/README.md
@@ -1,2 +1,11 @@
-![Under Construction](https://navtec.de/wp-content/uploads/2018/02/under-construction-2408061_1920.png)
+# Loesung 2
+
+Autonomes Fahrzeug 
+
+Ablauf:
+Fahrzeug fährt gerade aus solange kein Objekt sich näher befindet als GrenzeDistanz.
+Falls ein Objekt sich vor dem Fahrzeug befindet. Wird kontrolliert ob leicht links oder leicht rechts Ausweichmöglichkeiten vorhanden sind.
+Das Fahrzeug dreht sich dann ab , wo mehr Platz zur Verfügung steht.
+
+
